trt-1792: Add all regressions view to RegressedTestsModal.js
